backend/requirements.txt
fastapi 
uvicorn 
jinja2 
python-multipart
python-dotenv
firebase-admin
google-cloud-firestore